refactor: Fixes for privacy screen (#4114)
